LayoutTests/http/tests/security/window-opened-from-sandboxed-iframe-should-inherit-sandbox-expected.txt

Tests that windows created from a sandboxed context inherit the same sandbox

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframeContentWindow.didRunScriptInPopup is false
PASS successfullyParsed is true

TEST COMPLETE
